// src/types/indexer/indexerManual.ts
type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>;
export type PartialBy<T, K extends keyof T> = Omit<T, K> & Partial<Pick<T, K>>;

export interface IndexerCompositeFillResponse {
  pageSize?: number;
  totalResults?: number;
  offset?: number;
  fills?: IndexerCompositeFillObject[];
}

export interface IndexerCompositeOrderObject {
  id: string;
  subaccountId?: string;
  clientId?: string;
  clobPairId?: string;
  side: IndexerOrderSide;
  size: string;
  totalFilled: string;
  price: string;
  type: IndexerOrderType;
  reduceOnly?: boolean;
  orderFlags?: string;
  goodTilBlock?: string | null;
  goodTilBlockTime?: string | null;
  createdAtHeight?: string | null;
  clientMetadata?: string;
  triggerPrice?: string | null;
  timeInForce?: IndexerAPITimeInForce;
  status?: IndexerAPIOrderStatus;
  postOnly?: boolean;
  ticker: string;
  updatedAt?: IndexerIsoString | null;
  updatedAtHeight?: string | null;
  subaccountNumber: number;
  removalReason?: string;
  totalOptimisticFilled?: string;
}
// ...
export type IndexerWsOrderUpdate = Partial<
  Omit<IndexerCompositeOrderObject, 'subaccountNumber'>
> & { id: string };